Move code pieces to correct places
